hotfix(medusa-plugin-brightpearl): remove AUTH/CAPTURE
Removes the AUTH/CAPTURE flow; as it creates an issue when a user wants to
delete an authorized order. The order will at that point have a payment and
can thereby not be deleted. By simply using the RECEIPT payment type we avoid
this issue and make it possible to do extended authorization periods.